refactor(app_user): home → event cross-feature import 해소 (event_now_bar_controller)
Closes #851

matching data providers (matchCandidates, myMatches, myVoteCount,
myVotedCandidateIds, maxVoteCount) moved from event feature's
matching_vote_controller.dart to minglit_kit's matching_repository.dart.
These are thin wrappers around MatchingRepository — they belong with
the repository in the shared package.

MatchingVoteController (UI action logic) remains in event feature.

Constraint: Feature-first architecture requires no cross-feature imports
Rejected: Coordinator pattern | providers are data-only, not navigation
Confidence: high
Scope-risk: moderate — all existing callers already import minglit_kit
