+++ apps/app_user/lib/src/features/home/widgets/event_now_bar_controller.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 
-import 'package:app_user/src/features/event/matching/matching_vote_controller.dart';
 import 'package:minglit_kit/minglit_kit.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
