+++ apps/app_user/lib/src/features/home/widgets/event_now_bottom_sheet.dart
@@ -1,4 +1,3 @@
-import 'package:app_user/src/features/event/matching/matching_vote_controller.dart';
 import 'package:app_user/src/features/event/matching/widgets/matching_vote_content.dart';
 import 'package:app_user/src/features/home/widgets/event_now_bar_controller.dart';
 import 'package:app_user/src/features/ticket/data/ticket_wallet_repository.dart';
